+++ tests/Unit/Service/CraServiceTest.php
@@ -2,11 +2,13 @@
 
 namespace App\Tests\Unit\Service;
 
+use App\Entity\User;
 use App\Repository\CraRepository;
 use App\Repository\ProjetRepository;
 use App\Service\CraService;
 use App\Service\DateMonthService;
 use App\Service\JoursFeriesCalculator;
+use DateTime;
 use PHPUnit\Framework\TestCase;
 
 class CraServiceTest extends TestCase
@@ -52,5 +54,89 @@ class CraServiceTest extends TestCase
 
         $this->assertEquals('11', $cra->getMois()->format('m'));
         $this->assertEquals('01', $cra->getMois()->format('d'));
+    }
+
+    public function testCreateDefaultCraTakesAccountOfUserDateEntreeSameMonth()
+    {
+        $craService = new CraService(new DateMonthService(), $this->craRepositoryMock, $this->projetRepositoryMock, new JoursFeriesCalculator());
+
+        $user = new User();
+        $user->setDateEntree(DateTime::createFromFormat('Y-m-d', '2020-11-10'));
+
+        $cra = $craService->loadCraForUser($user, DateTime::createFromFormat('Y-m-d', '2020-11-01'));
+
+        $expected = [
+            0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 1, 0, 1, 1, 0, 0,
+            1, 1, 1, 1, 1, 0, 0,
+            1, 1, 1, 1, 1, 0, 0,
+            1,
+        ];
+
+        $this->assertEquals($expected, $cra->getJours());
+    }
+
+    public function testCreateDefaultCraTakesAccountOfUserDateEntreeLongBefore()
+    {
+        $craService = new CraService(new DateMonthService(), $this->craRepositoryMock, $this->projetRepositoryMock, new JoursFeriesCalculator());
+
+        $user = new User();
+        $user->setDateEntree(DateTime::createFromFormat('Y-m-d', '2020-11-09'));
+
+        $cra = $craService->loadCraForUser($user, DateTime::createFromFormat('Y-m-d', '2020-04-01'));
+
+        $expected = [
+            0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0,
+        ];
+
+        $this->assertEquals($expected, $cra->getJours());
+    }
+
+    public function testCreateDefaultCraTakesAccountOfUserDateSortieSameMonth()
+    {
+        $craService = new CraService(new DateMonthService(), $this->craRepositoryMock, $this->projetRepositoryMock, new JoursFeriesCalculator());
+
+        $user = new User();
+        $user->setDateSortie(DateTime::createFromFormat('Y-m-d', '2020-11-10'));
+
+        $cra = $craService->loadCraForUser($user, DateTime::createFromFormat('Y-m-d', '2020-11-01'));
+
+        $expected = [
+            0,
+            1, 1, 1, 1, 1, 0, 0,
+            1, 1, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0,
+        ];
+
+        $this->assertEquals($expected, $cra->getJours());
+    }
+
+    public function testCreateDefaultCraTakesAccountOfUserDateSortieLongAfter()
+    {
+        $craService = new CraService(new DateMonthService(), $this->craRepositoryMock, $this->projetRepositoryMock, new JoursFeriesCalculator());
+
+        $user = new User();
+        $user->setDateSortie(DateTime::createFromFormat('Y-m-d', '2020-05-10'));
+
+        $cra = $craService->loadCraForUser($user, DateTime::createFromFormat('Y-m-d', '2021-11-01'));
+
+        $expected = [
+            0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0,
+            0,
+        ];
+
+        $this->assertEquals($expected, $cra->getJours());
     }
 }