+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Tous les changements apportés au projet sont suivis dans ce fichier.
 Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0/).
 
 ## [À venir dans la prochaine version]
+### Ajouté
+- Ajout de la date d'entrée et de sortie d'un collaborateur
+    - L'admin peut définir les dates
+    - Saisie des absences : les jours où l'user n'est pas dans la société sont décochés par défaut
+    - Les feuilles de temps affichent 0 heures les jours où l'user n'est pas dans la société
+    - L'entrée et la sortie d'un user sont affichées dans son activité
 ### Réparé
 - Certaines années affichées (saisie temps/absences) n'étaient pas la bonne en janvier